fix: Set LTI 1.3 launch values after authorizing with Canvas
If a user needed to authorize with Canvas before using the application, the application would fail with an internal server error after authorizing. If you refreshed the page, the application would launch correctly but not immediately after authorizing.

This was happening because the authorization flow wasn't setting some necessary LTI 1.3 values.

This commit adds those values and the application runs correctly after authorizing with Canvas. No page refresh needed.