fix: Use client_id from params in OpenID initialization
We were previously pulling the `client_id` from the first `LtiInstall` record
with a matching `iss`. This approach only works if there's only one
`LtiInstall` record with a matching `iss` though. In practice, we may very
well have multiple `LtiInstall` records with the same `iss`.
The request parameters include the `client_id` for the tool being launched.
We can use this `client_id` instead of pulling it from the `LtiInstall` record
and then we don't have to worry about `iss` collisions.