feature: Make LtiDeployment a child of LtiInstall
We're making `LtiDeployment` a child of `LtiInstall` because it will give
us access to the `client_id` in a few different contexts where we need
it. Right now, we're finding the `client_id` by calling `application.client_id(iss)`.
This isn't a dependable way to get a `client_id` though because there
may be multiple `LtiInstall` records with the same `iss`. Associating
an `LtiDeployment` with an `LtiInstall` will give us a dependable way
to determine the correct client ID in those contexts where we need it.

Conceptually, making `LtiDeployment` a child of `LtiInstall` also makes
sense. We create an `LtiDeployment` based on a given `LtiInstall`.