fix: Update LtiInstall getter methods to return correct record
Previously, with only the `iss` to search with, these methods wouldn't
necessarily return the correct `LtiInstall` record. The `oidc_url`,
`token_url` and `jwks_url` attributes should be the same between
`LtiInstall` records with the same `iss`, so, in practice, they should still
return the correct values. However, the fact remains that we'll be pulling
(potentially) the wrong `LtiInstall` record and then assuming that that
wrong `LtiInstall` record has the same `oidc_url`, etc. as the correct
`LtiInstall` record. That might be a correct assumption, but it still feels
to me that it's going to lead to bugs. You think you're dealing with the
`oidc_url` attribute from the `LtiInstall` record associated with the
install you're working with but it's actually a different record. That's
confusing and surprising.

By updating these methods so they require the `client_id` attribute
along with the `iss`, we can be certain we're fetching the correct
`LtiInstall` record because there's a unique constraint in the database
that won't allow duplicate `iss`/`client_id` pairs.

Because of the reasons described above, it doesn't make sense to have
a `client_id` method. There's no way for an `Application` to determine
the correct `client_id`. That's why that method was removed. We're making
changes elsewhere in the codebase that will make the `client_id` accessible
where it's needed.