fix: Don't modify string in place
We were modifying, in place, the string we got from the `endpoint`
method. This meant that, if you called `endpoint` again after executing
`url << "?#{query}"` you'd get the modified string instead of the original
string `endpoint` returned. This wasn't causing any bugs given our
current implementation, but I find the behavior confusing and surprising,
so I think it makes sense to `dup` the string to avoid the confusion
and prevent potential future bugs.