tests: Update mock response data to include names
We updated the code to check for names in the response data from the
names and roles API to determine whether or not the LTI key associated
with the application was set to private.

We forgot to update the mock response data though, so a couple of tests
were failing.