fix: Eliminate invalid "X-Frame-Options" header
`ALLOWALL` isn't a valid directive for the `X-Frame-Options` header.
The possible directives are `DENY` and `SAMEORIGIN` [1]. Not adding
an `X-Frame-Options` header allows our application to be embedded
in any page. That should give us the behavior we want, and that will
eliminate the error being shown in the browser console.

[1]: https://developer.mozilla.org/en-US/docs/Web/HTTP/Headers/X-Frame-Options