refactor: Change throw to raise
I don't think we use `throw` anywhere else in our ruby code in this
application to create exceptions; we always use `raise`. `throw` and
`raise` do similar things, but they're not the same and you work with
them differently.

I think it would be easier to work with this error if this was a `raise`
instead of a `throw`.