fix: Change learner to user
We originally extracted this code from Atomic Journals. In the context
of Atomic Journals, it makes sense to call the users in the results "learners"
because we only fetched learners and that's what Facing History calls
the students in their courses.

In this more generic context though, "learners" doesn't really make sense.
We should call them something like "users".