Update oauth redirect url
Pull url out of params and redirect to that value.
We interpolate it to get brakeman to stop warning on it.
This oauth_complete_url value is constructed by us and stuck into the params.
So there is no security risk in redirecting to it.
Brakeman doesn't like redirecting to user input values.
As this doesn't come from user input, we will just pull the value out of
the params and redirect to it.