Correctly namespace concerns
From https://web.archive.org/web/20210216182939/https://guides.rubyonrails.org/upgrading_ruby_on_rails.html#concerns 
```
In that case, app/models/concerns is assumed to be a root directory (because it belongs to the autoload paths), and it is ignored as namespace. So, app/models/concerns/foo.rb should define Foo, not Concerns::Foo.

The Concerns:: namespace worked with the classic autoloader as a side-effect of the implementation, but it was not really an intended behavior. An application using Concerns:: needs to rename those classes and modules to be able to run in zeitwerk mode.
```