app_callback_url to JWT and pass the JWT in OAuth form
We can't trust app_callback_url in the client because it it used as the redirect in the OAuth process. Instead of trusting a form value we embed it in the JWT so that it is signed and cannot be modified by the client. Then before we redirect to Canvas to do OAuth we extract app_callback_url from the JWT and include it in the state that is stored in the database. This is then retrieved after returning from a successful OAuth and used as a redirect.